Added duration attributes and methods
I realize I'd forgotten to do this before; at the rate this project is going, they may never even be used, but it's better to be safe than to be sorry. I've also changed the constructor for the CommonTask to not include a completion; you can only decide the name, description, and duration; completion is always set to 0 at initialization.