Implemented toList methods
One thing that might be useful, especially for the purposes of presenting the Curriculum's TaskTree to the UI, would be maybe doing a hashmap where each key is a supertask ID, and each value is a Task object. I'll make a note of this and include it in my PR to bring attention to it.