Created more TaskTree tests; fixed a bug in updateTask
More importantly, there is now a small if statement that simply skips the updateTask() method if there is no assigned Task. I would like for this instead to return an exception, maybe TasklessTreeError or something. I also fixed a small bug in removeChildTaskTree() that the warnings made me learn about.