Fixed divide-by-zero bug in updateTask
I was notified that my implementation, as it was, would divide by zero if a TaskTree had its last task removed. I've included both a call to updateTask in the removeSubTaskTree method to assure TaskTrees are always updated, and a setting of an ignoring of the taskCount in the case that the last TaskTree was just removed. If the last TaskTree is removed, then at that point, the parent Tree would already have completion set only either to 0 or 100, so no adjustment of completion need be made.